+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.42.0]
+
+## Added
+- Support for create/delete replicas in the NoSQL Database service
+- Support for update/delete row in a table in NoSQL Database service
+
+## Changed
+- Minimum required [OCI Python SDK](https://github.com/oracle/oci-python-sdk) changed to 2.123.0
+
 ## [4.41.0]
 
 ## Added